# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Blog Personnel - ISEP_P5",
        "description": "Cette API permet la gestion d'un blog personnel : authentification par OTP, gestion des articles, des amis et des commentaires.",
        "contact": {
            "name": "Équipe ISEP",
            "email": "support@isep.sn"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local (dev)"
        },
        {
            "url": "https://api.monblog.isep.sn",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Récupérer les articles du fil d'actualité",
                "operationId": "11aa4cc6561a7b94c56d1201d62513ee",
                "responses": {
                    "200": {
                        "description": "Liste des articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "articles": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Créer un nouvel article",
                "operationId": "e328bed23caf889ad982ad3f1180bccb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Mon premier article"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Contenu de l'article..."
                                    },
                                    "is_public": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "allow_comments": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-01-01T12:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article créé avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/articles/my": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Récupérer mes articles",
                "operationId": "cc4cb7fbdf2eacc6bb37b44525d46f38",
                "responses": {
                    "200": {
                        "description": "Liste de mes articles"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Récupérer un article spécifique",
                "operationId": "b3e529eab04efb5d305b07e2e58d9869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article récupéré avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Modifier un article",
                "operationId": "2717d5a8e3b1f16637aded325347f9ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "is_public": {
                                        "type": "boolean"
                                    },
                                    "allow_comments": {
                                        "type": "boolean"
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article modifié avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Supprimer un article",
                "operationId": "97f34592611a1dd53a086865e29706c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article supprimé avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/articles/{article_id}/comments": {
            "post": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Ajouter un commentaire à un article",
                "operationId": "b13d9f63e09bb84721285d15a89a64c4",
                "parameters": [
                    {
                        "name": "article_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Excellent article !"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire ajouté avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/{id}": {
            "delete": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Supprimer un commentaire",
                "operationId": "721e5dc0c32ef81276203df965767606",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaire supprimé avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}